Split integ tests in A LOT of modules
