+++ tests/integration/test_charm.py
@@ -9,7 +9,6 @@ import re
 import typing
 from secrets import token_hex
 
-import magic
 import pytest
 import requests
 from juju.action import Action
@@ -18,10 +17,8 @@ from juju.model import Model
 from juju.unit import Unit
 from ops.model import ActiveStatus
 from pytest_operator.plugin import OpsTest
-from saml_test_helper import SamlK8sTestHelper
 
 import synapse
-from tests.integration.helpers import create_moderators_room
 
 # caused by pytest fixtures
 # pylint: disable=too-many-arguments
@@ -66,102 +63,6 @@ async def test_synapse_validate_configuration(synapse_app: Application):
     await synapse_app.model.wait_for_idle(
         idle_period=30, timeout=120, apps=[synapse_app.name], status="active"
     )
-
-
-@pytest.mark.cos
-@pytest.mark.usefixtures("synapse_app", "prometheus_app")
-async def test_prometheus_integration(
-    model: Model,
-    prometheus_app_name: str,
-    synapse_app_name: str,
-    get_unit_ips: typing.Callable[[str], typing.Awaitable[tuple[str, ...]]],
-):
-    """
-    arrange: after Synapse charm has been deployed.
-    act: establish relations established with prometheus charm.
-    assert: prometheus metrics endpoint for prometheus is active and prometheus has active scrape
-        targets.
-    """
-    await model.add_relation(prometheus_app_name, synapse_app_name)
-    await model.wait_for_idle(
-        idle_period=30, apps=[synapse_app_name, prometheus_app_name], status=ACTIVE_STATUS_NAME
-    )
-
-    for unit_ip in await get_unit_ips(prometheus_app_name):
-        query_targets = requests.get(f"http://{unit_ip}:9090/api/v1/targets", timeout=10).json()
-        assert len(query_targets["data"]["activeTargets"])
-
-
-@pytest.mark.cos
-@pytest.mark.usefixtures("synapse_app", "prometheus_app", "grafana_app")
-async def test_grafana_integration(
-    model: Model,
-    synapse_app_name: str,
-    prometheus_app_name: str,
-    grafana_app_name: str,
-    get_unit_ips: typing.Callable[[str], typing.Awaitable[tuple[str, ...]]],
-):
-    """
-    arrange: after Synapse charm has been deployed.
-    act: establish relations established with grafana charm.
-    assert: grafana Synapse dashboard can be found.
-    """
-    await model.relate(
-        f"{prometheus_app_name}:grafana-source", f"{grafana_app_name}:grafana-source"
-    )
-    await model.relate(synapse_app_name, grafana_app_name)
-
-    await model.wait_for_idle(
-        apps=[synapse_app_name, prometheus_app_name, grafana_app_name],
-        status=ACTIVE_STATUS_NAME,
-        idle_period=60,
-    )
-
-    action = await model.applications[grafana_app_name].units[0].run_action("get-admin-password")
-    await action.wait()
-    password = action.results["admin-password"]
-    grafana_ip = (await get_unit_ips(grafana_app_name))[0]
-    sess = requests.session()
-    sess.post(
-        f"http://{grafana_ip}:3000/login",
-        json={
-            "user": "admin",
-            "password": password,
-        },
-    ).raise_for_status()
-    datasources = sess.get(f"http://{grafana_ip}:3000/api/datasources", timeout=10).json()
-    datasource_types = set(datasource["type"] for datasource in datasources)
-    assert "prometheus" in datasource_types
-    dashboards = sess.get(
-        f"http://{grafana_ip}:3000/api/search",
-        timeout=10,
-        params={"query": "Synapse Operator"},
-    ).json()
-    assert len(dashboards)
-
-
-@pytest.mark.nginx
-@pytest.mark.usefixtures("synapse_app")
-async def test_nginx_route_integration(
-    model: Model,
-    nginx_integrator_app: Application,
-    synapse_app_name: str,
-    nginx_integrator_app_name: str,
-):
-    """
-    arrange: build and deploy the Synapse charm, and deploy the nginx-integrator.
-    act: relate the nginx-integrator charm with the Synapse charm.
-    assert: requesting the charm through nginx-integrator should return a correct response.
-    """
-    await model.add_relation(f"{synapse_app_name}", f"{nginx_integrator_app_name}")
-    await nginx_integrator_app.set_config({"service-hostname": synapse_app_name})
-    await model.wait_for_idle(idle_period=30, status=ACTIVE_STATUS_NAME)
-
-    response = requests.get(
-        "http://127.0.0.1/_matrix/static/", headers={"Host": synapse_app_name}, timeout=5
-    )
-    assert response.status_code == 200
-    assert "Welcome to the Matrix" in response.text
 
 
 async def test_reset_instance_action(
@@ -213,111 +114,6 @@ async def test_workload_version(
         version_match = re.search(synapse.SYNAPSE_VERSION_REGEX, server_version)
         assert version_match
         assert version_match.group(1) == juju_workload_version
-
-
-async def test_synapse_enable_mjolnir(
-    ops_test: OpsTest,
-    synapse_app: Application,
-    access_token: str,
-    get_unit_ips: typing.Callable[[str], typing.Awaitable[tuple[str, ...]]],
-):
-    """
-    arrange: build and deploy the Synapse charm, create an user, get the access token,
-        enable Mjolnir and create the management room.
-    act: check Mjolnir health point.
-    assert: the Synapse application is active and Mjolnir health point returns a correct response.
-    """
-    await synapse_app.set_config({"enable_mjolnir": "true"})
-    await synapse_app.model.wait_for_idle(
-        idle_period=30, timeout=120, apps=[synapse_app.name], status="blocked"
-    )
-    synapse_ip = (await get_unit_ips(synapse_app.name))[0]
-    create_moderators_room(synapse_ip, access_token)
-    async with ops_test.fast_forward():
-        # using fast_forward otherwise would wait for model config update-status-hook-interval
-        await synapse_app.model.wait_for_idle(
-            idle_period=30, apps=[synapse_app.name], status="active"
-        )
-
-    res = requests.get(f"http://{synapse_ip}:{synapse.MJOLNIR_HEALTH_PORT}/healthz", timeout=5)
-
-    assert res.status_code == 200
-
-
-@pytest.mark.nginx
-@pytest.mark.asyncio
-@pytest.mark.usefixtures("nginx_integrator_app")
-async def test_saml_auth(  # pylint: disable=too-many-locals
-    model: Model,
-    model_name: str,
-    synapse_app: Application,
-    get_unit_ips: typing.Callable[[str], typing.Awaitable[tuple[str, ...]]],
-):
-    """
-    arrange: integrate Synapse with SAML and NGINX charms, upload metadata to samltest.id
-        and configure public_baseurl.
-    act: simulate a user authenticating via SAML.
-    assert: The SAML authentication process is executed successfully.
-    """
-    synapse_config = await synapse_app.get_config()
-    server_name = synapse_config["server_name"]["value"]
-    assert server_name
-    saml_helper = SamlK8sTestHelper.deploy_saml_idp(model_name)
-
-    saml_integrator_app: Application = await model.deploy(
-        "saml-integrator",
-        channel="latest/edge",
-        series="jammy",
-        trust=True,
-    )
-    await model.wait_for_idle()
-    saml_helper.prepare_pod(model_name, f"{saml_integrator_app.name}-0")
-    saml_helper.prepare_pod(model_name, f"{synapse_app.name}-0")
-    await saml_integrator_app.set_config(
-        {
-            "entity_id": f"https://{saml_helper.SAML_HOST}/metadata",
-            "metadata_url": f"https://{saml_helper.SAML_HOST}/metadata",
-        }
-    )
-    await model.wait_for_idle(idle_period=30)
-    await model.add_relation(saml_integrator_app.name, synapse_app.name)
-    await model.wait_for_idle(
-        idle_period=30,
-        apps=[synapse_app.name, saml_integrator_app.name],
-        status=ACTIVE_STATUS_NAME,
-    )
-
-    session = requests.session()
-    headers = {"Host": server_name}
-    for unit_ip in await get_unit_ips(synapse_app.name):
-        response = session.get(
-            f"http://{unit_ip}:{synapse.SYNAPSE_NGINX_PORT}/_synapse/client/saml2/metadata.xml",
-            timeout=10,
-            headers=headers,
-        )
-        assert response.status_code == 200
-        saml_helper.register_service_provider(name=server_name, metadata=response.text)
-        saml_page_path = "_matrix/client/r0/login/sso/redirect/saml"
-        saml_page_params = "redirectUrl=http%3A%2F%2Flocalhost%2F&org.matrix.msc3824.action=login"
-        redirect_response = session.get(
-            f"http://{unit_ip}:{synapse.SYNAPSE_NGINX_PORT}/{saml_page_path}?{saml_page_params}",
-            verify=False,
-            timeout=10,
-            headers=headers,
-            allow_redirects=False,
-        )
-        assert redirect_response.status_code == 302
-        redirect_url = redirect_response.headers["Location"]
-        saml_response = saml_helper.redirect_sso_login(redirect_url)
-        assert f"https://{server_name}" in saml_response.url
-
-        url = saml_response.url.replace(
-            f"https://{server_name}", f"http://{unit_ip}:{synapse.SYNAPSE_NGINX_PORT}"
-        )
-        logged_in_page = session.post(url, data=saml_response.data, headers={"Host": server_name})
-
-        assert logged_in_page.status_code == 200
-        assert "Continue to your account" in logged_in_page.text
 
 
 @pytest.mark.parametrize(
@@ -496,146 +292,24 @@ async def test_anonymize_user(
     assert res.status_code == 403
 
 
-@pytest.mark.s3
-@pytest.mark.usefixtures("s3_backup_bucket")
-async def test_synapse_enable_s3_backup_integration_success(
+@pytest.mark.usefixtures("synapse_app")
+async def test_nginx_route_integration(
     model: Model,
-    synapse_app: Application,
-    s3_integrator_app_backup: Application,
+    nginx_integrator_app: Application,
+    synapse_app_name: str,
+    nginx_integrator_app_name: str,
 ):
     """
-    arrange: Synapse App deployed and s3-integrator deployed with bucket created.
-    act:  integrate s3-integrator with Synapse.
-    assert: Synapse gets into active status.
+    arrange: build and deploy the Synapse charm, and deploy the nginx-integrator.
+    act: relate the nginx-integrator charm with the Synapse charm.
+    assert: requesting the charm through nginx-integrator should return a correct response.
     """
-    await model.add_relation(s3_integrator_app_backup.name, f"{synapse_app.name}:backup")
-    await model.wait_for_idle(apps=[s3_integrator_app_backup.name], status=ACTIVE_STATUS_NAME)
+    await model.add_relation(f"{synapse_app_name}", f"{nginx_integrator_app_name}")
+    await nginx_integrator_app.set_config({"service-hostname": synapse_app_name})
+    await model.wait_for_idle(idle_period=30, status=ACTIVE_STATUS_NAME)
 
-    await model.wait_for_idle(
-        idle_period=30,
-        apps=[synapse_app.name, s3_integrator_app_backup.name],
-        status=ACTIVE_STATUS_NAME,
+    response = requests.get(
+        "http://127.0.0.1/_matrix/static/", headers={"Host": synapse_app_name}, timeout=5
     )
-
-
-async def test_synapse_enable_s3_backup_integration_no_bucket(
-    model: Model,
-    synapse_app: Application,
-    s3_integrator_app_backup: Application,
-):
-    """
-    arrange: Synapse App deployed and s3-integrator deployed.
-    act:  integrate s3-integrator with Synapse.
-    assert: Synapse gets into blocked status because the bucket does not exist.
-    """
-    await model.add_relation(s3_integrator_app_backup.name, f"{synapse_app.name}:backup")
-    await model.wait_for_idle(apps=[s3_integrator_app_backup.name], status=ACTIVE_STATUS_NAME)
-
-    await model.wait_for_idle(apps=[synapse_app.name], idle_period=5, status="blocked")
-    assert synapse_app.units[0].workload_status == "blocked"
-    assert "bucket does not exist" in synapse_app.units[0].workload_status_message
-
-
-@pytest.mark.s3
-@pytest.mark.usefixtures("s3_backup_bucket")
-async def test_synapse_create_backup_correct(
-    model: Model,
-    synapse_app: Application,
-    s3_integrator_app_backup: Application,
-    s3_backup_configuration: dict,
-    boto_s3_client: typing.Any,
-):
-    """
-    arrange: Synapse App deployed and related with s3-integrator. backup_passphrase set.
-    act: Run create-backup action
-    assert: Correct response from the action that includes the backup-id.
-       An encrypted object was created in S3 with the correct name.
-    """
-    await model.add_relation(s3_integrator_app_backup.name, f"{synapse_app.name}:backup")
-    passphrase = token_hex(16)
-    await synapse_app.set_config({"backup_passphrase": passphrase})
-    await model.wait_for_idle(
-        idle_period=30,
-        apps=[synapse_app.name, s3_integrator_app_backup.name],
-        status=ACTIVE_STATUS_NAME,
-    )
-
-    synapse_unit: Unit = next(iter(synapse_app.units))
-    backup_action: Action = await synapse_unit.run_action("create-backup")
-    await backup_action.wait()
-
-    assert backup_action.status == "completed"
-    assert "backup-id" in backup_action.results
-    bucket_name = s3_backup_configuration["bucket"]
-    path = s3_backup_configuration["path"].strip("/")
-    object_key = f"{path}/{backup_action.results['backup-id']}"
-    s3objresp = boto_s3_client.get_object(Bucket=bucket_name, Key=object_key)
-    objbuf = s3objresp["Body"].read()
-    assert "GPG symmetrically encrypted data (AES256 cipher)" in magic.from_buffer(objbuf)
-
-
-@pytest.mark.s3
-@pytest.mark.usefixtures("s3_backup_bucket")
-async def test_synapse_create_backup_no_passphrase(
-    model: Model,
-    synapse_app: Application,
-    s3_integrator_app_backup: Application,
-):
-    """
-    arrange: Synapse App deployed and related with s3-integrator. No backup_passphrase.
-    act: Run create-backup action
-    assert: The action fails because there is no passphrase.
-    """
-    await synapse_app.reset_config(["backup_passphrase"])
-    await model.add_relation(s3_integrator_app_backup.name, f"{synapse_app.name}:backup")
-    await model.wait_for_idle(
-        idle_period=30,
-        apps=[synapse_app.name, s3_integrator_app_backup.name],
-        status=ACTIVE_STATUS_NAME,
-    )
-
-    synapse_unit: Unit = next(iter(synapse_app.units))
-    backup_action: Action = await synapse_unit.run_action("create-backup")
-    await backup_action.wait()
-
-    assert backup_action.status == "failed"
-    assert "backup-id" not in backup_action.results
-    assert "Missing backup_passphrase" in backup_action.message
-
-
-@pytest.mark.s3
-@pytest.mark.usefixtures("s3_backup_bucket")
-async def test_synapse_list_backups(
-    model: Model,
-    synapse_app: Application,
-    s3_integrator_app_backup: Application,
-):
-    """
-    arrange: Synapse App deployed and related with s3-integrator. Set backup_passphrase
-        and create two backups.
-    act: Run action list-backups
-    assert: There should be two backups, with the same keys as the ones created.
-    """
-    await model.add_relation(s3_integrator_app_backup.name, f"{synapse_app.name}:backup")
-    passphrase = token_hex(16)
-    await synapse_app.set_config({"backup_passphrase": passphrase})
-    await model.wait_for_idle(
-        idle_period=30,
-        apps=[synapse_app.name, s3_integrator_app_backup.name],
-        status=ACTIVE_STATUS_NAME,
-    )
-    synapse_unit: Unit = next(iter(synapse_app.units))
-    backup_action_1: Action = await synapse_unit.run_action("create-backup")
-    await backup_action_1.wait()
-    backup_action_2: Action = await synapse_unit.run_action("create-backup")
-    await backup_action_2.wait()
-
-    list_backups_action: Action = await synapse_unit.run_action("list-backups")
-    await list_backups_action.wait()
-
-    assert list_backups_action.status == "completed"
-    assert "backups" in list_backups_action.results
-    backups = list_backups_action.results["backups"]
-    assert len(backups) == 2
-    assert backup_action_1.results["backup-id"] in backups
-    assert backup_action_2.results["backup-id"] in backups
+    assert response.status_code == 200
+    assert "Welcome to the Matrix" in response.text
